+++ back-end/api/index.js
@@ -1,0 +1,3 @@
+import { author } from "./author.js";
+
+export { author };