+++ back-end/utils/utils.js
@@ -1,0 +1,7 @@
+export const FormateData = (data) => {
+    if (data) {
+      return { data };
+    } else {
+      throw new Error("Data Not found!");
+    }
+  };